// src/app/shared/models/AssetTrackingModel.ts
import { AssetModel } from "./AssetModel";
import { EmployeeModel } from "./EmployeeModel";

export class AssetTrackingModel {

    AssetTrackingId?: number;
    From?: string;
    To?: string;
    astId?: number;
    usrId?: number;
    DateTime?: Date;
    usrAccountName?: string;
    EmployeeName?: string;
 //   Asset?: AssetModel;

  }